+++ lib/collection.dart
@@ -1,7 +1,7 @@
 library more.collection;
 
-export 'src/collection/bimap.dart' show BiMap;
-export 'src/collection/bitlist.dart' show BitList;
+export 'src/collection/bimap.dart' show BiMap, BiMapExtension;
+export 'src/collection/bitlist.dart' show BitList, BitListExtension;
 export 'src/collection/double_range.dart'
     show DoubleRange, DoubleRangeExtension;
 export 'src/collection/integer_range.dart'