+++ lib/iterable.dart
@@ -9,7 +9,7 @@ export 'src/iterable/group.dart' show GroupExtension, Group;
 export 'src/iterable/indexed.dart' show IndexedExtension, Indexed;
 export 'src/iterable/iterate.dart' show iterate;
 export 'src/iterable/periodical.dart'
-    show Period, PeriodicalDateTime, TruncateDateTime;
+    show Period, PeriodicalExtension, TruncateExtension;
 export 'src/iterable/permutations.dart' show PermutationsExtension;
 export 'src/iterable/product.dart' show ProductExtension;
 export 'src/iterable/repeat.dart' show repeat;